pyproject.toml
[project]
name = "mlops-lab11"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.30.0",
    "joblib>=1.5.3",
    "pydantic-settings>=2.12.0",
    "skl2onnx>=1.19.1",
]

[dependency-groups]
deployment = [
    "boto3>=1.35.0",
    "joblib>=1.4.0",
    "onnxscript>=0.5.7",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "scikit-learn>=1.5.0",
]
inference = [
    "awslambdaric>=3.1.1",
    "onnxruntime>=1.23.2",
    "tokenizers>=0.22.2",
    "mangum>=0.20.0",
]
integration = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
